requirements.txt
aiohttp==3.8.5
beautifulsoup4==4.9.3
boto3==1.34.27
click==8.1.7
kafka_python==2.0.2
python-dotenv==1.0.1
pytz==2023.3.post1
Requests==2.31.0
setuptools==69.0.2
setuptools==65.5.0
urllib3==1.26.18
xmltodict==0.13.0
